Super simplifying

+++ OpenPics/Classes/Providers/DPLA/OPDPLAProvider.m
@@ -119,7 +119,7 @@ NSString * const OPProviderTypeDPLA = @"com.saygoodnight.dpla";
             returnCanLoadMore = YES;
         }
         
-        NSLog(@"LAST: %d   TOTAL: %d", lastItem, totalItems);
+        NSLog(@"LAST: %ld   TOTAL: %ld", (long)lastItem, (long)totalItems);
         if (success) {
             success(retArray, returnCanLoadMore);
         }
